+++ samples/IfItQuacks.Sample.Methods/Program.cs
@@ -1,0 +1,4 @@
+using IfItQuacks.Sample.Methods;
+
+Ops.Foo(new A());
+Ops.Foo(new B());