+++ samples/IfItQuacks.Sample.Properties/Program.cs
@@ -1,0 +1,4 @@
+using IfItQuacks.Sample.Properties;
+
+var person = new Person { Name = "Steven" };
+Ops.Greet(person);